+++ include/iceflow/consumer.hpp
@@ -33,7 +33,7 @@ class IceflowConsumer {
 public:
   IceflowConsumer(std::shared_ptr<IceFlow> iceflow, const std::string &subTopic,
                   // TODO: Change constructor parameter to std::set as well.
-                  const std::vector<int> &topicPartitions)
+                  const std::vector<uint64_t> &topicPartitions)
       : m_iceflow(iceflow), m_subTopic(subTopic)
 
   {
@@ -45,15 +45,15 @@ public:
 
   std::vector<uint8_t> receiveData() { return m_inputQueue.waitAndPopValue(); }
 
-  void setTopicPartitions(const std::vector<int> &topicPartitions) {
+  void setTopicPartitions(const std::vector<uint64_t> &topicPartitions) {
     if (topicPartitions.empty()) {
       throw std::invalid_argument(
           "At least one topic partition has to be defined!");
     }
 
     if (auto validIceflow = m_iceflow.lock()) {
-      std::unordered_set<int> topicPartitionSet(topicPartitions.begin(),
-                                                topicPartitions.end());
+      std::unordered_set<uint64_t> topicPartitionSet(topicPartitions.begin(),
+                                                     topicPartitions.end());
 
       std::vector<decltype(m_subscriptionHandles)::key_type> handlesToErase;
       for (auto subscriptionHandle : m_subscriptionHandles) {
@@ -81,7 +81,7 @@ public:
   }
 
 private:
-  uint32_t subscribeToTopicPartition(int topicPartition) {
+  uint32_t subscribeToTopicPartition(uint64_t topicPartition) {
     if (auto validIceflow = m_iceflow.lock()) {
 
       std::function<void(std::vector<uint8_t>)> pushDataCallback =
@@ -108,7 +108,7 @@ private:
   const std::weak_ptr<IceFlow> m_iceflow;
   const std::string m_subTopic;
 
-  std::unordered_map<int, uint32_t> m_subscriptionHandles;
+  std::unordered_map<uint64_t, uint32_t> m_subscriptionHandles;
 
   RingBuffer<std::vector<uint8_t>> m_inputQueue;
 };
